Adding optional Rust backend.
If the BITSTRING_USE_RUST_CORE environment variable is set when the module is imported it will use bitformat as a backend instead of bitarray. This will only work for Python 3.11 or later.

Not all Rust core tests are passing, but equally they're not yet being run in the CI.